+++ src/main/kotlin/calculation/Difference.kt
@@ -1,0 +1,9 @@
+package calculation
+
+class Difference(lhs: String, rhs: String) : MathNumber() {
+    override val value: Number
+
+    init {
+        value = lhs.toInt() - rhs.toInt()
+    }
+}
