+++ src/main/kotlin/calculation/Product.kt
@@ -1,0 +1,9 @@
+package calculation
+
+class Product(lhs: String, rhs: String) : MathNumber() {
+    override val value: Number
+
+    init {
+        value = lhs.toInt() * rhs.toInt()
+    }
+}
