+++ src/main/kotlin/calculation/Sum.kt
@@ -1,0 +1,9 @@
+package calculation
+
+class Sum(lhs: String, rhs: String) : MathNumber() {
+    override val value: Number
+
+    init {
+        value = lhs.toInt() + rhs.toInt()
+    }
+}
